edited some stuff, separated the purchase order into service and api
