+++ frontend/lib/models/ui/purchase_order.dart
@@ -1,134 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:http/http.dart' as http;
+import 'package:frontend/models/api/purchase_order.dart';
+import 'package:frontend/services/purchase_order_service.dart';
 import 'dart:convert';
-
-// Exception for network errors
-class NetworkException implements Exception {
-  final String message;
-  NetworkException(this.message);
-  @override
-  String toString() => 'NetworkException: $message';
-}
-
-// Exception for data parsing errors
-class DataParsingException implements Exception {
-  final String message;
-  DataParsingException(this.message);
-  @override
-  String toString() => 'DataParsingException: $message';
-}
-
-class PurchaseOrder {
-  String purchaseOrderCode;
-  String? purchaseOrderFile;
-  String? suppliersPackingList;
-  int quantityPurchased;
-  DateTime orderDate;
-  DateTime expectedDeliveryDate;
-  double cost;
-
-  PurchaseOrder({
-    required this.purchaseOrderCode,
-    this.purchaseOrderFile,
-    this.suppliersPackingList,
-    required this.quantityPurchased,
-    required this.orderDate,
-    required this.expectedDeliveryDate,
-    required this.cost,
-  });
-
-  factory PurchaseOrder.fromJson(Map<String, dynamic> json) {
-    try {
-      return PurchaseOrder(
-        purchaseOrderCode: json['purchaseOrderCode'],
-        purchaseOrderFile: json['purchaseOrderFile'],
-        suppliersPackingList: json['suppliersPackingList'],
-        quantityPurchased: json['quantityPurchased'],
-        orderDate: DateTime.parse(json['orderDate']),
-        expectedDeliveryDate: DateTime.parse(json['expectedDeliveryDate']),
-        cost: (json['cost'] as num).toDouble(),
-      );
-    } catch (e) {
-      throw DataParsingException('Error parsing PurchaseOrder from JSON: $e');
-    }
-  }
-
-  Map<String, dynamic> toJson() => {
-    'purchaseOrderCode': purchaseOrderCode,
-    'purchaseOrderFile': purchaseOrderFile,
-    'suppliersPackingList': suppliersPackingList,
-    'quantityPurchased': quantityPurchased,
-    'orderDate': orderDate.toIso8601String(),
-    'expectedDeliveryDate': expectedDeliveryDate.toIso8601String(),
-    'cost': cost,
-  };
-}
-
-class PurchaseOrderService {
-  static const String baseUrl = 'http://localhost:8080/po/v1';
-
-  Future<List<PurchaseOrder>> fetchPurchaseOrders() async {
-    try {
-      final response = await http.get(Uri.parse('$baseUrl/getPOs'));
-      if (response.statusCode == 200) {
-        final List data = json.decode(response.body);
-        return data.map((json) => PurchaseOrder.fromJson(json)).toList();
-      } else {
-        throw NetworkException('Failed to load purchase orders: ${response.statusCode}');
-      }
-    } catch (e) {
-      if (e is FormatException) {
-        throw DataParsingException('Invalid JSON response: $e');
-      }
-      rethrow; // Rethrow other exceptions like NetworkException or SocketException
-    }
-  }
-
-  Future<void> addPurchaseOrder(PurchaseOrder po) async {
-    try {
-      final response = await http.post(
-        Uri.parse('$baseUrl/addPO'),
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode(po.toJson()),
-      );
-      if (response.statusCode != 200 && response.statusCode != 201) {
-        throw NetworkException('Failed to add purchase order: ${response.statusCode}');
-      }
-    } catch (e) {
-      rethrow;
-    }
-  }
-
-  Future<void> updatePurchaseOrder(PurchaseOrder po) async {
-    try {
-      final response = await http.put(
-        Uri.parse('$baseUrl/updatePO'),
-        headers: {'Content-Type': 'application/json'},
-        body: json.encode(po.toJson()),
-      );
-      if (response.statusCode != 200) {
-        throw NetworkException('Failed to update purchase order: ${response.statusCode}');
-      }
-    } catch (e) {
-      rethrow;
-    }
-  }
-
-  Future<void> deletePurchaseOrder(String purchaseOrderCode) async {
-    try {
-      final response = await http.delete(
-        Uri.parse('$baseUrl/deletePO/$purchaseOrderCode'),
-        headers: {'Content-Type': 'application/json'},
-      );
-      if (response.statusCode != 200) {
-        throw NetworkException('Failed to delete purchase order: ${response.statusCode}');
-      }
-    } catch (e) {
-      rethrow;
-    }
-  }
-}
 
 class PurchaseOrderPage extends StatefulWidget {
   const PurchaseOrderPage({super.key});
